fastem gui controller: refactor and move fastem specific controllers to own module
* rename alignmentController to calibration controller
* rename previous calibrationcontroller to calibrationsregionscontroller
